Move TransactionStatusBadge to common
